+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -313,6 +313,10 @@ Changes in existing checks
 
   - Do not report explicit call to destructor after move as an invalid use.
 
+- Improved :doc:`container-data-pointer
+  <clang-tidy/checks/readability/container-data-pointer>` check to use
+  ``c_str()`` for ``const``'s when available and present on a container.
+
 - Improved :doc:`cppcoreguidelines-avoid-capturing-lambda-coroutines
   <clang-tidy/checks/cppcoreguidelines/avoid-capturing-lambda-coroutines>`
   check by adding the `AllowExplicitObjectParameters` option. When enabled,